Updated release download link
I updated the release link and unified the version, since it is the same for all OS.

BTW, maybe is better provide the link do the releases page.

And... only for windows we have a "Setup" word. Maybe we should name "Altus-(windows|linux|macOS)". At macOS is kind of a setup too.